% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_lab_data.R
\name{associate_labs_with_outcomes}
\alias{associate_labs_with_outcomes}
\title{expects output from attach_covariates_outcomes}
\usage{
associate_labs_with_outcomes(
  labs,
  covariates = covariate_list_generator(),
  type = "prognostic",
  num_cores = 30
)
}
\description{
expects output from attach_covariates_outcomes
}
